+++ docs/source/history.rst
@@ -28,7 +28,7 @@ Unreleased.  Release 0.61.0
 - Remove support to use `~xoeuf.models.base.get_modelname`:func: with model
   proxies.
 
-- Depreciate model proxies `xoeuf.models.proxy`:mod:.
+- Deprecate model proxies `xoeuf.models.proxy`:mod:.
 
 
 2019-05-03.  Release 0.60.0